third_party/blink/web_tests/platform/mac-mac11-arm64/virtual/fenced-frame-mparch/wpt_internal/fenced_frame/visual-viewport.https-expected.txt

This is a testharness.js-based test.
FAIL visualViewport values inside fenced frame promise_test: Unhandled rejection with value: object "Error: assert_equals: [PRECONDITION] document scrolled x expected 30 but got 0"
Harness: the test ran to completion.
